Couette and MD coupled
In this commit the Couette code run on 1 processor and it can be
coupled with an arbitrary number of MD processors.  The code is still
at the hack level, but is does the main coupling operation.

It run ok for 100 continuum steps if the constrain force is not activated.

To get the executables instrumented with the coupler use the task
"couette_md" for  coupler/src_code/makefile.

Couette code is in the new folder Couette_serial.

In this commit the coupler algorithm has been changed significantly.
Detailed documentation to follow. The main steps of the algorithm are as follow:

1) MD uses its processor topology and CFD global grid details to compute
   the rage of CFD indices that cover the local domain.

2) MD and CFD exchange data and compute which section of the covering
   grid belongs to which CFD rank. (This is a map).

3) Using the maps data are exchanged at each iteration between CFD and MD:
   i) CFD sends velocities for the continuum correction to the force,
  ii) MD sends boundary local averages for the CFD boundary condition.

Also steps were done to disentangle the coupler subroutine from the
solver subroutines. In the initial version coupler subroutines were
accessing data from either code via the "USE" statement. This is not
going to work easily if the continuum or the MD code are replaced.

In the new version at the intiliasation stage all the data needed by
the coupler subroutines ( e.g. grid sizes, number of processors) are
copied in a coupler module. All the other coupler subroutines use this
module and argument variables to do their job.

So far this transformation was done only on the continuum side, MD will
follow shortly.

git-svn-id: http://svn.ma.ic.ac.uk/subversion/edward/MDNS_repo/branch/Couette_serial@21 1e92d776-e0e5-45cd-b670-8958d04d46c8