added constrain dynamic in y directions
Periodic boundary condition removed in MPI topology
along y direction.

Molecules are kept in the MD region by the repulsive force at the top
and the fixes molecules at the bottom.

This version does not has a proper correction to density needed to
counterbalance the influence of the constrain force at the top.
As a hack I left the top slab of MD cell empty at the initialisation stage.

The continuum grid dimension and other parameters are set to correspond to
the first test in Nie's paper.

git-svn-id: http://svn.ma.ic.ac.uk/subversion/edward/MDNS_repo/branch@28 1e92d776-e0e5-45cd-b670-8958d04d46c8