added dynamic number of processors for parallel runs
No need to recompile if the number of processor is changed.
However if the calculation is a restart one needs to  redistribute the
molecules (not implemented yet in parallel_io.mpi2.f90)

nproc is taken from MPI_Comm_size, its value is set  by the argument of mpirun.

The static arrays form modules whose sizes depend on nproc are
declared allocatable and allocated in messenger_init.

The topology of the processor grid (npx,npy,npz) is set in MD.in
The new keywork PROCESSORS was added in the locate subroutine.

A test is made in messenger_init for nproc == npx*npy*npz

git-svn-id: http://svn.ma.ic.ac.uk/subversion/edward/MDNS_repo/branch/MD_dCSE@56 1e92d776-e0e5-45cd-b670-8958d04d46c8