coupler input mechanism
From this version common parameters can be specified in
COUPLER.in file. The settings are distributed to the realms
in the initialisation stage.

A sample COUPLER.in may be found in coupler_dCSE/src_code/.

Some flexibility is provided, if a parameter is not specified
in COUPLER.in it will be read from the realm input file and
adjusted if necessary (with a log message).

In documentation directory the coupler_flow_chart.tex describes
in some detail the input operations.

This commit is just a functional prototype, more refinements are
to follow.

git-svn-id: http://svn.ma.ic.ac.uk/subversion/edward/MDNS_repo/branch/coupler_dCSE@116 1e92d776-e0e5-45cd-b670-8958d04d46c8