Top makefile added
coupled or uncoupled builds can be produced with the following command:

make BUILD=<build> PLATFORM=<platform> <target>

for a list of available targets type make help

Other changes :

VERSION make variable replaced with BUILD ( in line with common terminology)
make.inc was renamed to  platforms
dummy modules removed in coupler and Couette directories.

<platform>.inc file reorganised:
The compiler flag variable start from a common base and extra bit are added for specific
application with the help of if constructs.

Compiler flags are set according to the application name, the branch construct was removed,
some sanity checks were added in the .inc file, to be sure that the intended build is made.

git-svn-id: http://svn.ma.ic.ac.uk/subversion/edward/MDNS_repo/branch/MD_dCSE@128 1e92d776-e0e5-45cd-b670-8958d04d46c8