Corrections and reorganisation of continuum and constrain force calculation
constrain and continuum force are computed in MD (not in coupler as before)
using MD computational cells for efficiency ( to be measured!)

some small cosmetic changes/comments made in coupler
tests passed

more refinment is needed for continuum force:

This version neglects the molecules that are outside the local MD
domain. Statistically their impact should be small.
Future version should communicate to the corresponding rank the sums of velocities and
acceleration for molecules outside local MD domain.

git-svn-id: http://svn.ma.ic.ac.uk/subversion/edward/MDNS_repo/branch/MD_dCSE@162 1e92d776-e0e5-45cd-b670-8958d04d46c8