extended hector platforms make include files along intel one
Also added in DNS lib_src makefile flags
for cray and nag compiler.
HECToR lib files go in lib/$(PLATFORM).inc, useful
for testing. One has to set this manually when compiling the
library.

Code names in CFD makefiles were changed to couette and
serial_couette (serial is a  particular case useful
for testing/debbuging). The new names are in agreement with the target
name form the top makefile.

git-svn-id: http://svn.ma.ic.ac.uk/subversion/edward/MDNS_repo/branch/Couette_serial@183 1e92d776-e0e5-45cd-b670-8958d04d46c8