added debug subroutines which can dump uc in a given xy slice
The data is dumped by CGD and MD in separated files in
a format suitable for gnuplot.
This is work in progress but I though I better put it here before it
is left too much behind.

Anyway this data dumping is activated only if  the preprocessor flag
COUPLER_DEBUG_LA is set. Currently is activated only in hector-gnu
platform

One more change, in md.f90 the order call of continnum force and
send MD averages subrotuines was reversed to be consistent with
CFD.

git-svn-id: http://svn.ma.ic.ac.uk/subversion/edward/MDNS_repo/branch/coupler_dCSE@247 1e92d776-e0e5-45cd-b670-8958d04d46c8