Started adding cell by cell boundary application to continuum_coupler_socket. Currently each processor interpolates and extrapolates with no exchnaged between - this should be improved to exchange halos (for each cell) and interpolate on all.
Removed addition of nh from md_coupler_socket during cumulative velocity calculation. 
This is NOT correct as it ibins does not include halo cells so bottom cell of domain 
passed as empty! This bug was not noticed previously because cells were missed in both 
sum of molecules and velocity so the average still works out correctly.

socket_get_domain_top has been improved to check type of coupling and removed molecules
for OCT/NCER but not off/flekkoy...

Corrected Globalise routine in coupler

Added word 'printf' to routine


git-svn-id: http://svn.ma.ic.ac.uk/subversion/edward/MDNS_repo/branch/MD_dCSE@446 1e92d776-e0e5-45cd-b670-8958d04d46c8