Renamed a few folders within the codes DT and ES will be maintaining to be more in line with standard layouts. For example, the coupler library is no longer built in cpl-library/lib, but in cpl-library/cpl/lib for portability.
I've adjusted the makefiles in transflow, cpl, couette-serial and flowmol to
take into account the new location of the platforms directory, which is now in
misc/platforms, alongside misc/documentation and misc/inputs, which are other
folders that I thought cluttered the repository and didn't know where to put.
If you've got any objections to that let me know, I'm happy to move them back.

Also, I cleaned up the top-level makefile, so that's much much easier to follow
now.