1. Updated utils/postproclib/openfoamrawdata.py to be more careful with reading lists - the user can now select a specific entry in an output file by name with OpenFOAM_RawData.read_list_named_entry().
2. md_coupler_socket velocity send subroutine now only allocates one cell in
the y-direction (the halo cell). This routine is in a state of flux and we
need to work out what the best thing to do is. Should we allocate only one
cell for sending, or do we generalise for multiple cells and the halo/no halo
choice?

3. cplIcoFoam socket seems to receive the MD data correctly now. The unpacking
doesn't yet work with the boundary field patch, where strange behaviour occurs
with fvPatchFields/receiveVelocity/receiveVelocityFvPatchVectorField. This must
be fixed.