1. Updated the README to be more README-like, and moved its old contents to a more-appropriately-named file called GIT.
2. Simplified the build process for coupled programs. The misc/platforms
directory still exists for backwards compatibility, but its functionality has
been stripped back somewhat to be replaced by a newer directory called "make".
Inside make/ there are two files that are included from the makefiles in
cpl/src and lammps/src, one is reminiscent of the old "PLATFORM_default"
functionality, and the other is a template.inc "platform" for a new user to
simply define their own compilers and flags for debug and opt builds.

The flowmol makefile had too much going on for me to replace it now, so it
just uses the old system with misc/platforms. Updating it is on the todo list.